[refactor]판매 등록폼 관리 리팩토링 #68
